+++ aquadoggo/src/lib.rs
@@ -32,7 +32,7 @@ mod test_utils;
 #[cfg(test)]
 mod tests;
 
-pub use crate::api::LockFile;
+pub use crate::api::{ConfigFile, LockFile};
 pub use crate::config::{AllowList, Configuration};
 pub use crate::network::NetworkConfiguration;
 pub use node::Node;