+++ aquadoggo_cli/src/config.rs
@@ -1,30 +1,23 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use std::convert::TryFrom;
 use std::net::SocketAddr;
 use std::path::PathBuf;
-use std::str::FromStr;
-use std::sync::OnceLock;
 
-use anyhow::{anyhow, bail, Result};
-use aquadoggo::{AllowList, Configuration as NodeConfiguration, NetworkConfiguration};
+use anyhow::{bail, Result};
+use aquadoggo::{AllowList, ConfigFile, Configuration};
 use clap::{crate_version, Parser};
 use colored::Colorize;
 use directories::ProjectDirs;
 use figment::providers::{Env, Format, Serialized, Toml};
 use figment::Figment;
 use libp2p::PeerId;
-use p2panda_rs::schema::SchemaId;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use tempfile::TempDir;
+use serde::{Serialize, Serializer};
 
 use crate::utils::absolute_path;
 
 const WILDCARD: &str = "*";
 
 const CONFIG_FILE_NAME: &str = "config.toml";
-
-static TMP_DIR: OnceLock<TempDir> = OnceLock::new();
 
 type ConfigFilePath = Option<PathBuf>;
 
@@ -34,7 +27,7 @@ type ConfigFilePath = Option<PathBuf>;
 ///
 /// Returns a partly unchecked configuration object which results from all of these sources. It
 /// still needs to be converted for aquadoggo as it might still contain invalid values.
-pub fn load_config() -> Result<(ConfigFilePath, Configuration)> {
+pub fn load_config() -> Result<(ConfigFilePath, ConfigFile)> {
     // Parse command line arguments first to get optional config file path
     let cli = Cli::parse();
 
@@ -50,7 +43,7 @@ pub fn load_config() -> Result<(ConfigFilePath, Configuration)> {
         None => try_determine_config_file_path(),
     };
 
-    let mut figment = Figment::from(Serialized::defaults(Configuration::default()));
+    let mut figment = Figment::from(Serialized::defaults(ConfigFile::default()));
     if let Some(path) = &config_file_path {
         figment = figment.merge(Toml::file(path));
     }
@@ -270,138 +263,6 @@ where
     }
 }
 
-/// Configuration derived from environment variables and .toml file.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct Configuration {
-    pub log_level: String,
-    pub allow_schema_ids: UncheckedAllowList,
-    pub database_url: String,
-    pub database_max_connections: u32,
-    pub http_port: u16,
-    pub quic_port: u16,
-    pub blobs_base_path: Option<PathBuf>,
-    pub private_key: Option<PathBuf>,
-    pub mdns: bool,
-    pub direct_node_addresses: Vec<SocketAddr>,
-    pub allow_peer_ids: UncheckedAllowList,
-    pub block_peer_ids: Vec<PeerId>,
-    pub relay_addresses: Vec<SocketAddr>,
-    pub relay_mode: bool,
-    pub worker_pool_size: u32,
-}
-
-impl Default for Configuration {
-    fn default() -> Self {
-        let database_url = {
-            // Give each in-memory SQLite database an unique name as we're observing funny issues with
-            // SQLite sharing data between processes (!) and breaking each others databases
-            // potentially.
-            //
-            // See related issue: https://github.com/p2panda/aquadoggo/issues/568
-            let db_name = format!("dbmem{}", rand::random::<u32>());
-
-            // Set "mode=memory" to enable SQLite running in-memory and set "cache=shared", as
-            // setting it to "private" would break sqlx / SQLite.
-            //
-            // See related issue: https://github.com/launchbadge/sqlx/issues/2510
-            format!("sqlite://file:{db_name}?mode=memory&cache=shared")
-        };
-
-        Self {
-            log_level: "off".into(),
-            allow_schema_ids: UncheckedAllowList::Wildcard,
-            database_url,
-            database_max_connections: 32,
-            http_port: 2020,
-            quic_port: 2022,
-            blobs_base_path: None,
-            mdns: true,
-            private_key: None,
-            direct_node_addresses: vec![],
-            allow_peer_ids: UncheckedAllowList::Wildcard,
-            block_peer_ids: Vec::new(),
-            relay_addresses: vec![],
-            relay_mode: false,
-            worker_pool_size: 16,
-        }
-    }
-}
-
-impl TryFrom<Configuration> for NodeConfiguration {
-    type Error = anyhow::Error;
-
-    fn try_from(value: Configuration) -> Result<Self, Self::Error> {
-        // Check if given schema ids are valid
-        let allow_schema_ids = match value.allow_schema_ids {
-            UncheckedAllowList::Wildcard => AllowList::<SchemaId>::Wildcard,
-            UncheckedAllowList::Set(str_values) => {
-                let schema_ids: Result<Vec<SchemaId>, anyhow::Error> = str_values
-                    .iter()
-                    .map(|str_value| {
-                        SchemaId::from_str(str_value).map_err(|_| {
-                            anyhow!(
-                                "Invalid schema id '{str_value}' found in 'allow_schema_ids' list"
-                            )
-                        })
-                    })
-                    .collect();
-
-                AllowList::Set(schema_ids?)
-            }
-        };
-
-        // Check if given peer ids are valid
-        let allow_peer_ids = match value.allow_peer_ids {
-            UncheckedAllowList::Wildcard => AllowList::<PeerId>::Wildcard,
-            UncheckedAllowList::Set(str_values) => {
-                let peer_ids: Result<Vec<PeerId>, anyhow::Error> = str_values
-                    .iter()
-                    .map(|str_value| {
-                        PeerId::from_str(str_value).map_err(|_| {
-                            anyhow!("Invalid peer id '{str_value}' found in 'allow_peer_ids' list")
-                        })
-                    })
-                    .collect();
-
-                AllowList::Set(peer_ids?)
-            }
-        };
-
-        // Create a temporary blobs directory when none was given
-        let blobs_base_path = match value.blobs_base_path {
-            Some(path) => path,
-            None => TMP_DIR
-                .get_or_init(|| {
-                    // Initialise a `TempDir` instance globally to make sure it does not run out of
-                    // scope and gets deleted before the end of the application runtime
-                    tempfile::TempDir::new()
-                        .expect("Could not create temporary directory to store blobs")
-                })
-                .path()
-                .to_path_buf(),
-        };
-
-        Ok(NodeConfiguration {
-            allow_schema_ids,
-            database_url: value.database_url,
-            database_max_connections: value.database_max_connections,
-            http_port: value.http_port,
-            blobs_base_path,
-            worker_pool_size: value.worker_pool_size,
-            network: NetworkConfiguration {
-                quic_port: value.quic_port,
-                mdns: value.mdns,
-                direct_node_addresses: value.direct_node_addresses,
-                allow_peer_ids,
-                block_peer_ids: value.block_peer_ids,
-                relay_addresses: value.relay_addresses,
-                relay_mode: value.relay_mode,
-                ..Default::default()
-            },
-        })
-    }
-}
-
 fn try_determine_config_file_path() -> Option<PathBuf> {
     // Find config file in current folder
     let mut current_dir = std::env::current_dir().expect("Could not determine current directory");
@@ -423,7 +284,7 @@ fn try_determine_config_file_path() -> Option<PathBuf> {
 pub fn print_config(
     private_key_path: Option<&PathBuf>,
     config_file_path: ConfigFilePath,
-    config: &NodeConfiguration,
+    config: &Configuration,
 ) -> String {
     println!(
         r"                       ██████ ███████ ████
@@ -530,52 +391,4 @@ Node is ready!
         private_key.blue(),
         relay_mode.blue(),
     )
-}
-
-/// Helper struct to deserialize from either a wildcard string "*" or a list of string values.
-///
-/// These string values are not checked yet and need to be validated in a succeeding step.
-#[derive(Debug, Clone)]
-pub enum UncheckedAllowList {
-    Wildcard,
-    Set(Vec<String>),
-}
-
-impl Serialize for UncheckedAllowList {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        match self {
-            UncheckedAllowList::Wildcard => serializer.serialize_str(WILDCARD),
-            UncheckedAllowList::Set(list) => list.serialize(serializer),
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for UncheckedAllowList {
-    fn deserialize<D>(deserializer: D) -> Result<UncheckedAllowList, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(untagged)]
-        enum Value<T> {
-            String(String),
-            Vec(Vec<T>),
-        }
-
-        let value = Value::deserialize(deserializer)?;
-
-        match value {
-            Value::String(str_value) => {
-                if str_value == WILDCARD {
-                    Ok(UncheckedAllowList::Wildcard)
-                } else {
-                    Err(serde::de::Error::custom("only wildcard strings allowed"))
-                }
-            }
-            Value::Vec(vec) => Ok(UncheckedAllowList::Set(vec)),
-        }
-    }
 }